Refactor CLI
This commit refactors the CLI code to have each command in a separate
file. It also cleans up the CLI interface to be more consistent and
intuitive.
